Issue#186 : updated take(until:) description and added unit test.
Prior to this change `take(until:)` was described as terminating the receiver only for a completed event on the given signal. This has been updated to accurately state that it is for _any_ event that was sent on the provided signal.

Unit Tests provided and full unit test suite passed.