Fixed deadlock in retry signal
In rare cases, there is a race condition when a retrying signal is disposed while the subject it is listening to is sending an event. This was causing a deadlock between the `dispose()` and the `on(_:)` functions of the observer.

This fixes it. I’ve tested tons of different scenarios. I have not been able to get this test case to pass consistently with any other method I’ve tried. With this method, my test passes consistently with any number of iterations, retries, and intervals.

The test case:
https://github.com/AnthonyMDev/ReactiveKitDeadlockDemo